+++ launcher/LauncherIcon.cpp
@@ -1177,7 +1177,7 @@ LauncherIcon::OnRemoteCountVisibleChanged(LauncherEntryRemote* remote)
 {
   if (remote->CountVisible())
   {
-    SetEmblemText(std::to_string((long long)remote->Count()));
+    SetEmblemText(std::to_string(remote->Count()));
   }
   else
   {
